feat: add base refactoring
